requirements.txt
py-cord~=2.6.0
pytest~=8.3.2
pytest-cov~=5.0.0
python-dotenv~=1.0.0
slack_sdk~=3.32.0